pyproject.toml
[tool.poetry]
name = "bb-cli"
version = "0.1.0"
description = ""
authors = ["Johnatas <johnatasr@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.1,<4.0"
black = "^23.12.0"
pytest = "^7.4.3"
isort = "^5.13.2"
requests = "^2.31.0"
click = "^8.1.7"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
